+++ src/chevron_blue/renderer.py
@@ -1,9 +1,15 @@
+from __future__ import annotations
+
 import io
-import sys
+import typing as t
+import warnings
 from collections.abc import Callable, Iterator, Sequence
-from os import linesep, path
+from os import path
 
 from .tokenizer import tokenize
+
+if t.TYPE_CHECKING:
+    OnMissingKey = t.Literal["ignore", "warn", "error"]
 
 #
 # Helper functions
@@ -26,7 +32,7 @@ def _html_escape(string):
     return string
 
 
-def _get_key(key, scopes, warn, keep, def_ldel, def_rdel):
+def _get_key(key, scopes, on_missing_key: OnMissingKey, keep, def_ldel, def_rdel):
     """Get a key from the current scope"""
 
     # If the key is a dot
@@ -73,8 +79,14 @@ def _get_key(key, scopes, warn, keep, def_ldel, def_rdel):
 
     # We couldn't find the key in any of the scopes
 
-    if warn:
-        sys.stderr.write("Could not find key '%s'%s" % (key, linesep))
+    if on_missing_key == "warn":
+        warnings.warn(
+            "Could not find key '%s' in data" % key,
+            UserWarning,
+            stacklevel=2,
+        )
+    elif on_missing_key == "error":
+        raise KeyError("Could not find key '%s'" % key)
 
     if keep:
         return "%s %s %s" % (def_ldel, key, def_rdel)
@@ -121,9 +133,10 @@ def render(
     def_ldel="{{",
     def_rdel="}}",
     scopes=None,
-    warn=False,
+    warn=None,
     keep=False,
     no_escape=False,
+    on_missing_key: OnMissingKey | None = None,
 ):
     """Render a mustache template.
 
@@ -169,17 +182,34 @@ def render(
 
     scopes        -- The list of scopes that get_key will look through
 
-    warn          -- Issue a warning to stderr when a template substitution isn't found in the data
+    warn          -- When true, issue warnings.
+                     Deprecated; use `on_missing_key` instead. Equivalent to `on_missing_key="warn"`.
 
     no_escape     -- Do not HTML escape variable values
 
     keep          -- Keep unreplaced tags when a template substitution isn't found in the data
 
+    on_missing_key    -- How strict to be when a template substitution isn't found in the data.
+                     Can be one of:
+                     * permissive -- Do not warn or raise an exception
+                     * warn       -- Issue a warning to stderr
+                     * strict     -- Raise a KeyError
+                     (default: permissive)
 
     Returns:
 
     A string containing the rendered template.
     """
+
+    if warn is None:
+        on_missing_key = on_missing_key or "ignore"
+    else:
+        # If incompatible options are used, error
+        if on_missing_key is not None:
+            raise ValueError(
+                "The `warn` argument cannot be used with `on_missing_key`."
+            )
+        on_missing_key = "warn" if warn else "ignore"
 
     # If the template is a seqeuence but not derived from a string
     if isinstance(template, Sequence) and not isinstance(template, str):
@@ -224,7 +254,12 @@ def render(
         elif tag == "variable":
             # Add the html escaped key to the output
             thing = _get_key(
-                key, scopes, warn=warn, keep=keep, def_ldel=def_ldel, def_rdel=def_rdel
+                key,
+                scopes,
+                on_missing_key=on_missing_key,
+                keep=keep,
+                def_ldel=def_ldel,
+                def_rdel=def_rdel,
             )
             if thing is True and key == ".":
                 # if we've coerced into a boolean by accident
@@ -239,7 +274,12 @@ def render(
         elif tag == "no escape":
             # Just lookup the key and add it
             thing = _get_key(
-                key, scopes, warn=warn, keep=keep, def_ldel=def_ldel, def_rdel=def_rdel
+                key,
+                scopes,
+                on_missing_key=on_missing_key,
+                keep=keep,
+                def_ldel=def_ldel,
+                def_rdel=def_rdel,
             )
             if not isinstance(thing, str):
                 thing = str(thing)
@@ -249,7 +289,12 @@ def render(
         elif tag == "section":
             # Get the sections scope
             scope = _get_key(
-                key, scopes, warn=warn, keep=keep, def_ldel=def_ldel, def_rdel=def_rdel
+                key,
+                scopes,
+                on_missing_key=on_missing_key,
+                keep=keep,
+                def_ldel=def_ldel,
+                def_rdel=def_rdel,
             )
 
             # If the scope is a callable (as described in
@@ -299,7 +344,7 @@ def render(
                         def_ldel=def_ldel,
                         def_rdel=def_rdel,
                         scopes=data and [data] + scopes or scopes,
-                        warn=warn,
+                        on_missing_key=on_missing_key,
                         keep=keep,
                         no_escape=no_escape,
                     ),
@@ -339,7 +384,7 @@ def render(
                         partials_dict=partials_dict,
                         def_ldel=def_ldel,
                         def_rdel=def_rdel,
-                        warn=warn,
+                        on_missing_key=on_missing_key,
                         keep=keep,
                         no_escape=no_escape,
                     )
@@ -353,7 +398,12 @@ def render(
         elif tag == "inverted section":
             # Add the flipped scope to the scopes
             scope = _get_key(
-                key, scopes, warn=warn, keep=keep, def_ldel=def_ldel, def_rdel=def_rdel
+                key,
+                scopes,
+                on_missing_key=on_missing_key,
+                keep=keep,
+                def_ldel=def_ldel,
+                def_rdel=def_rdel,
             )
             scopes.insert(0, not scope)
 
@@ -378,7 +428,7 @@ def render(
                 def_rdel=def_rdel,
                 padding=part_padding,
                 scopes=scopes,
-                warn=warn,
+                on_missing_key=on_missing_key,
                 keep=keep,
                 no_escape=no_escape,
             )