+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+
+- Added `on_missing_key = "ignore" | "warn" | "error"` as an alternative to `warn` that errors on missing keys (see [#14](https://github.com/zanieb/chevron-blue/pull/14))
+
 ## 0.2.1
 
 - Fixed bug where `--no-escape` was not applied to recursive patterns; (see [#5](https://github.com/zanieb/chevron-blue/pull/5))